working on sim tests
